Revert IOUtils changes
