--- apm-agent-core/src/main/resources/META-INF/services/co.elastic.apm.agent.sdk.internal.util.IOUtils$IOUtilsProvider
@@ -1,1 +1,0 @@
-co.elastic.apm.agent.util.IOUtilsProviderImpl
